Backend refactor (#50)
