+++ server/helpers/database.js
@@ -1,0 +1,2 @@
+const Datastore = require("nedb-promises");
+module.exports = Datastore.create("database.db");
